+++ udata_front/theme/gouvfr/assets/js/types.ts
@@ -71,9 +71,30 @@ export type NewDiscussion = {
   subject: Subject;
 };
 
-export type CreateDiscussion = (discussion: NewDiscussion) => Promise<any>
+export type CreateDiscussion = (discussion: NewDiscussion) => Promise<any>;
 
-export type CreateComment = (comment: string) => Promise<any>
+export type CreateComment = (comment: string) => Promise<any>;
+
+export type Spam = {
+  status?: string;
+}
+
+export type Discussion = Array<{content: string, posted_by: User, posted_on: string, spam?: Spam}>;
+
+export type Thread = {
+  id: string;
+  discussion: Discussion;
+  title: string;
+  url: string;
+  closed: string;
+  closed_by: User;
+  spam?: Spam;
+};
+
+export type Sort = {
+  name: string,
+  key: string
+}
 
 export type Quality = {
   all_resources_available: boolean;
@@ -125,6 +146,10 @@ slug: string;
 quality: Quality;
 metrics: { discussions: number; followers: number; reuses: number; views: number; };
 harvest: Harvest | null;
+};
+
+export type UiDataset = Omit<Dataset, 'last_modified'> & {
+  last_modified: Date;
 };
 
 export type Reuse = Owned & {