+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Current (in progress)
 
+- Move Pagination to datagouv-components [#365](https://github.com/etalab/udata-front/pull/365)
+- Fix release CI [#380](https://github.com/etalab/udata-front/pull/380)
+- Add stories to Resource component [#364](https://github.com/etalab/udata-front/pull/364)
 - Move Well to datagouv-components [#382](https://github.com/etalab/udata-front/pull/382)
 
 ## 3.5.4 (2024-03-20)